Implimentation of enum classes for FS and tracking which model is being used
Strings to track constants leads to unreadable code and easy mistakes. Added enums to track these numbers to allow the code to be easier to read and better to edit.
Also added a switch-case (match-case in py) so its not a if statement chain which is horrible to read aswell